% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberFormatters.R
\name{format_p_ke}
\alias{format_p_ke}
\title{format_p_ke}
\usage{
format_p_ke(p, sci = TRUE, digits = 3, sci_thres = 0.001, ...)
}
\arguments{
\item{p}{numeric atom}

\item{sci}{use scientific x10 format}

\item{digits}{number of decimal places to display}

\item{sci_thres}{threshold below which to apply scientific format [0,1]}

\item{...}{parameters for kableExtra::cell_spec(). e.g. bold = TRUE}
}
\value{
character atom
}
\description{
a wrapper for kableExtra::cell_spec() around format_p_md
}
\examples{
format_p_ke(p = 0.1234, sci = TRUE, digits = 3, sci_thres = 1e-3, bold = TRUE)
}
