% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phdfunctions.R
\docType{package}
\name{phdFunctions}
\alias{phdFunctions}
\alias{phdFunctions-package}
\title{phdFunctions: a package with helpful formatting functions used in my thesis.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Formatting functions}{

The foo functions ...
}

