% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textConverters.R
\name{stat_to_short}
\alias{stat_to_short}
\title{stat_to_short}
\usage{
stat_to_short(x)
}
\description{
replaces the long statname with the abbrevation that can then be fed into short_to_latex
which creates the latex glossary command for the stat
}
