% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberFormatters.R
\name{three_dp}
\alias{three_dp}
\title{rounds a number to 3 decimal places and returns a string
convience function}
\usage{
three_dp(num)
}
\description{
rounds a number to 3 decimal places and returns a string
convience function
}
