% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberFormatters.R
\name{round_dp}
\alias{round_dp}
\title{round_dp}
\usage{
round_dp(num, dp)
}
\arguments{
\item{num}{a number}

\item{dp}{a whole number >= 0}
}
\value{
character string version of the float to specified decimal places
}
\description{
create a string formatted to n decimal places
}
